[WIP] chore: 実行スクリプトの場所を移動, `yarn run builder` を修正. classファイル化の途中.
