+++ test/dev/components/Atoms/vue-atom.vue
@@ -1,0 +1,20 @@
+<template>
+  <div class="test">test</div>
+</template>
+
+<script>
+export default {
+  props: {
+    isTesting: Boolean
+  },
+  data() {
+    return { debug: true }
+  }
+}
+</script>
+
+<style scoped>
+.test {
+  border: 1px solid #000;
+}
+</style>
