+++ test/dev/components/Molecules/vue-molecules.vue
@@ -1,0 +1,19 @@
+<template>
+  <div class="molecules">
+    test
+  </div>
+</template>
+
+<script>
+export default {
+  data: function() {
+    return { foo: 'bar' }
+  }
+}
+</script>
+
+<style scoped>
+.molecules {
+  font-size: 1rem;
+}
+</style>
